Make the stat pages suck less
Because we were using UIKit tabs; Svelte needed to render out the entire StatTableSet; which when using the "Select All" button, was very, very slow.

This introduces a (less pretty) tab component that allows Svelte to track Data/DOM separately and thus optimize, removing essentially all of the noticeable delay.